feat: 헬스체크 추가. SQL 어노테이션 native query로 변경
Former-commit-id: 51c4f7d2ff5e2ba8c6479b58b0bd1e7a8c9b2405

+++ back/src/main/java/gg/garen/back/duoRecommendation/service/DuoRecommendationService.java
@@ -41,7 +41,7 @@ public class DuoRecommendationService {
         for(DuoRecord duoRecord : duoRecords){
             //포지션별로 3개씩만 받게하기 고민할 것.
             String position = duoRecord.getLane1().equals(curPosition)? duoRecord.getLane2() : duoRecord.getLane1();
-            int championKey =  duoRecord.getChampion1() == curChampionKey? duoRecord.getChampion2() : duoRecord.getChampion1();
+            Long championKey =  duoRecord.getChampion1() == curChampionKey? duoRecord.getChampion2() : duoRecord.getChampion1();
             String champion = championKeyNameMap.get(championKey);
             int positionNum = -1;
             for(int i = 0; i < 5; i++){
